+++ APlayer/Converters/IFF-8SVX/IFF-8SVXConverter.cpp
@@ -876,7 +876,7 @@ ap_result IFF8SVXConverter::SaveHeader(PFile *file, const APConverter_SampleForm
 		stereoFile->Close();
 
 		// Build a new file name
-		newName = file->GetFullPath() + "aps" + PString::CreateHexNumber((uint32)stereoFile);
+		newName = file->GetFullPath() + "aps" + PString::CreateHexNumber((uintptr_t)stereoFile);
 		stereoFile->Open(newName, PFile::pModeReadWrite | PFile::pModeCreate);
 	}
